Componentize `App`
Componentize `App` to separate concerns.
